+++ receiver/prometheusreceiver/targetallocator/manager_test.go
