+++ receiver/prometheusreceiver/targetallocator/testdata/config.yaml
